requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Data Processing
pandas==2.1.3
pydantic==2.5.0

# Database
sqlalchemy==2.0.23

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# CLI
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
pathvalidate==3.3.1
slugify==8.0.1